+++ haskell/src/Neversoft/Audio.hs
@@ -23,7 +23,7 @@ import           Data.Bits
 import qualified Data.ByteString              as B
 import qualified Data.ByteString.Lazy         as BL
 import qualified Data.Conduit.Audio           as CA
-import           Data.Int
+import           Data.Int                     (Int64)
 import           Data.Maybe                   (fromMaybe)
 import           Data.SimpleHandle            (byteStringSimpleHandle,
                                                makeHandle)
@@ -33,6 +33,7 @@ import qualified Data.Text.Encoding           as TE
 import           Data.Word
 import           FFMPEG                       (ffSource)
 import           Neversoft.Checksum
+import           Sound.FSB
 import           System.FilePath              (takeFileName)
 
 ghworKeys :: [B.ByteString]
@@ -324,8 +325,8 @@ ghworEncrypt bs = do
   key <- makeKey $ ghworKeys !! fromIntegral keyIndex
   return $ ctrCombine key iv bs <> ghworCipher
 
-readFSB :: (MonadResource m) => BL.ByteString -> IO (CA.AudioSource m Int16)
-readFSB dec = let
+readFSBXMA :: (MonadResource m) => BL.ByteString -> IO (CA.AudioSource m Float)
+readFSBXMA dec = let
   dec' = BL.concat
     [ BL.take 0x62 dec
     , BL.singleton 0x10 -- this is a hack so ffmpeg recognizes the fsb's xma encoding. TODO send a PR to fix
@@ -333,6 +334,35 @@ readFSB dec = let
     ]
   readable = makeHandle "decoded FSB audio" $ byteStringSimpleHandle dec'
   in ffSource $ Left readable
+
+splitInterleavedMP3 :: Int64 -> BL.ByteString -> [BL.ByteString]
+splitInterleavedMP3 1 bs = [bs]
+splitInterleavedMP3 n bs = let
+  frameSize = 384 -- TODO actually figure this out, might vary.
+  -- see http://www.datavoyage.com/mpgscript/mpeghdr.htm "How to calculate frame length"
+  getPart i = BL.concat $ takeWhile (not . BL.null) $ do
+    j <- [i * frameSize, (i + n) * frameSize ..]
+    return $ BL.take frameSize $ BL.drop j bs
+  in map getPart [0 .. n - 1]
+
+splitFSBMP3 :: BL.ByteString -> IO [BL.ByteString]
+splitFSBMP3 dec = do
+  fsb <- parseFSB dec
+  let song = head $ either fsb3Songs fsb4Songs $ fsbHeader fsb
+  case quotRem (fromIntegral $ fsbSongChannels song) 2 of
+    (pairs, 0) -> return $ splitInterleavedMP3 pairs $ head $ fsbSongData fsb
+    _ -> fail $ "FSB MP3 claims to have odd number of channels (" <> show (fsbSongChannels song) <> ")"
+
+readFSB :: (MonadResource m) => BL.ByteString -> IO (CA.AudioSource m Float)
+readFSB dec = do
+  fsb <- parseFSB dec
+  let song = head $ either fsb3Songs fsb4Songs $ fsbHeader fsb
+  case fsbExtra song of
+    Left _xma -> readFSBXMA dec
+    Right _mp3 -> do
+      mp3s <- splitFSBMP3 dec
+      srcs <- forM mp3s $ \mp3 -> ffSource $ Left $ makeHandle "decoded FSB audio" $ byteStringSimpleHandle mp3
+      return $ foldl1 CA.merge srcs
 
 decryptFSB' :: FilePath -> B.ByteString -> Maybe BL.ByteString
 decryptFSB' name enc = let
@@ -348,7 +378,7 @@ convertAudio :: FilePath -> FilePath -> IO ()
 convertAudio fin fout = do
   dec <- decryptFSB fin >>= maybe (error "Couldn't decrypt GH .fsb.xen audio") return
   src <- readFSB dec
-  audioIO Nothing (CA.mapSamples CA.fractionalSample src) fout
+  audioIO Nothing src fout
 
 reverseMapping :: B.ByteString
 reverseMapping = B.pack $ flip map [0 .. 0xFF] $ \i -> foldr (.|.) 0